\name{BBMV-package}
\alias{BBMV-package}
\alias{BBMV}
\docType{package}
\title{
\packageTitle{BBMV}
}
\description{
\packageDescription{BBMV}
}
\details{
\packageDESCRIPTION{BBMV}
\packageIndices{BBMV}
}
\author{
\packageAuthor{BBMV}
Maintainer: \packageMaintainer{BBMV}
}
\references{
Inferring bounded evolution in phenotypic characters from phylogenetic comparative data. F.C. Boucher and V. Demery. Systematic Biology, 65, 651-661, 2016

A general model for estimating macroevolutionary landscapes from phylogenetic comparative data. F.C. Boucher, V. Demery, E. Conti, L. J. Harmon and J. Uyeda. Submitted
}

\keyword{ package }
